requirements.txt
annotated-types
blinker
boto3
botocore
certifi
cffi
charset-normalizer
click
crypto
cryptography
Flask
Flask-Cors
idna
itsdangerous
Jinja2
jmespath
logzero
MarkupSafe
Naked
numpy
pandas
psycopg2
pycparser
pycryptodome
pydantic
pydantic_core
PyJWT
python-dateutil
pytz
PyYAML
requests
s3transfer
schedule
shareconnect
shellescape
six
smartapi-python
typing_extensions
tzdata
urllib3
watchtower
websocket-client
Werkzeug
